# Simplify generation of castling moves