This is an untested non-functional simplification.  You should all be proud of me that I didn't test this (it was hard, but I resisted).  I can't imagine there is any performance difference.  Let me know if a test is required.