@mcostalba 2019-02-06
This code path is very hot, so yes please give it a try with perft because it should be slower. I wrote the current from after many trials and it was the fastest I came up with.