@protonspring 2019-02-06
Removing to check at the first if will probably be slower, because otherwise, we always have to check both sides.

I'll do some perf testing, but for some reason, my machine doesn't reflect what happens on the framework.