@protonspring 2019-02-06
This version is fastest on my machine, but the difference is barely perceptible ( probably just noise).
bench 16 1 7 current perft:
Master: 19026
Initial patch: 19030
final patch: 19017