@protonspring 2019-02-06
I tested more this morning for a few hours.  In very very long tests this patch 1.3% faster than master:
(./stockfish bench 16 1 8 current perft   -   repeating 10 times for each version)

master ave nps: 158042193
patch ave nps: 160174172

Can someone else verify?