@mcostalba 2019-02-07
@protonspring I guess you tested on **starting position** when castling is possible and pos.castling_impeded() is always true, in this case you skip a castling check condition, namely pos.can_castle(CastlingRight(OO | OOO)) that is always true.

Your test is interesting because we can infer a couple of things:

1) This code path is really hot, because just a very simple extra condition gives a measurable difference in perft speed.

2) The test is biased toward position with castling flags enabled, but this is not the common case, actually in real games, most of the time castling flag is not set.

Eventually a more neutral test could be:

`bench 16 1 5 default perft:`

Or even better, directly testing in a middle game position (set a position with UCI and then run bench on the new current one), with castling not set. In this case you can see if there is any regression in the common case.

If you have some time to spend and you are curious, it would be interesting to post the disassembled code of both versions. I don't think it is really worth it, it is just an exercise for the interested one to eventually learn something new. If I have some time (hardly) I will do myself, just out of curiosity.