@joergoster 2019-02-07
For some reason I get a tremendous speed boost in perft.
For comparison, I did 2 runs (perft 5) over default bench positions, 2 runs (perft 5, perft 6) over 100 random positions (opening, midgame, endgame), and finally a default bench up to depth 23.

current master first
```
SF-Master (profile-build x86-64-modern)
=========

bench 16 1 5 default perft

===========================
Total time (ms) : 13839
Nodes searched  : 1450447254
Nodes/second    : 104808675

===========================
Total time (ms) : 13561
Nodes searched  : 1450447254
Nodes/second    : 106957249


bench 16 1 5 100.epd perft

===========================
Total time (ms) : 59546
Nodes searched  : 6235674607
Nodes/second    : 104720293

bench 32 1 6 100.epd perft

===========================
Total time (ms) : 2284326
Nodes searched  : 241431253267
Nodes/second    : 105690366


bench 128 1 23 default depth

===========================
Total time (ms) : 131796
Nodes searched  : 175986206
Nodes/second    : 1335292
```

Now this patch
```
SF-Patch (protonspring) (profile-build x86-64-modern)
=======================

bench 16 1 5 default perft

===========================
Total time (ms) : 11309
Nodes searched  : 1450447254
Nodes/second    : 128256013

===========================
Total time (ms) : 11311
Nodes searched  : 1450447254
Nodes/second    : 128233335


bench 16 1 5 100.epd perft

===========================
Total time (ms) : 49168
Nodes searched  : 6235674607
Nodes/second    : 126823840


bench 32 1 6 100.epd perft

===========================
Total time (ms) : 1900796
Nodes searched  : 241431253267
Nodes/second    : 127015867


bench 128 1 23 default depth

===========================
Total time (ms) : 132150
Nodes searched  : 175986206
Nodes/second    : 1331715
```

At default bench searching up to depth 23, no big difference. But a tremendous speedup of about **17 %** when running perft!
Any explanation for this? No castling rights available for most of the positions?

This is the code I tested, slightly changed to adapt to SF's general code style (curly braces, blanks around 'OO, OOO'):
```
        if (Type != CAPTURES)
        {
            for (CastlingRight cr : { OO, OOO })
                if (!pos.castling_impeded(cr) && pos.can_castle(cr))
                    *moveList++ = make<CASTLING>(ksq, pos.castling_rook_square(cr));
        }
```
_Hardware: AMD FX(tm)-8120 Eight-Core Processor × 8
Software: Version Linux Mint 19.1 Tessa 64-bit, Kernel Linux 4.15.0-45-generic x86_64, MATE 1.20.1
gcc version 7.3.0 (Ubuntu 7.3.0-27ubuntu1~18.04)_